+++ backend/src/main/java/com/carffeine/carffeine/car/controller/CarController.java
@@ -4,8 +4,8 @@ import com.carffeine.carffeine.auth.controller.support.AuthMember;
 import com.carffeine.carffeine.car.controller.dto.MemberCarInfoResponse;
 import com.carffeine.carffeine.car.domain.Car;
 import com.carffeine.carffeine.car.domain.MemberCar;
-import com.carffeine.carffeine.car.infrastructure.dto.CarResponse;
-import com.carffeine.carffeine.car.infrastructure.dto.CarsResponse;
+import com.carffeine.carffeine.car.infrastructure.repository.dto.CarResponse;
+import com.carffeine.carffeine.car.infrastructure.repository.dto.CarsResponse;
 import com.carffeine.carffeine.car.service.CarQueryService;
 import com.carffeine.carffeine.car.service.CarService;
 import com.carffeine.carffeine.car.service.dto.CarRequest;
