+++ backend/src/main/java/com/carffeine/carffeine/car/controller/dto/MemberCarInfoResponse.java
@@ -1,7 +1,7 @@
 package com.carffeine.carffeine.car.controller.dto;
 
 import com.carffeine.carffeine.car.domain.MemberCar;
-import com.carffeine.carffeine.car.infrastructure.dto.CarResponse;
+import com.carffeine.carffeine.car.infrastructure.repository.dto.CarResponse;
 
 public record MemberCarInfoResponse(
         Long memberId,
