+++ backend/src/main/java/com/carffeine/carffeine/car/service/CarQueryService.java
@@ -1,7 +1,7 @@
 package com.carffeine.carffeine.car.service;
 
-import com.carffeine.carffeine.car.infrastructure.dto.CarsResponse;
 import com.carffeine.carffeine.car.infrastructure.repository.CarQueryRepository;
+import com.carffeine.carffeine.car.infrastructure.repository.dto.CarsResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
