+++ backend/src/test/java/com/carffeine/carffeine/car/service/CarQueryServiceTest.java
@@ -2,7 +2,7 @@ package com.carffeine.carffeine.car.service;
 
 import com.carffeine.carffeine.car.domain.Car;
 import com.carffeine.carffeine.car.domain.CarRepository;
-import com.carffeine.carffeine.car.infrastructure.dto.CarsResponse;
+import com.carffeine.carffeine.car.infrastructure.repository.dto.CarsResponse;
 import com.carffeine.carffeine.helper.integration.IntegrationTest;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
